Add header files basics.h and observables.h.
Basics.h contains definitions of structs for units and time.

Observables.h defines an abstract class Observable that contains units and
the time at which the observable was recorded. Derived template classes are
ScalarObservable and VectorObservable.

TODO: add comments to the code according to the
[Google Style Guide](https://google.github.io/styleguide/cppguide.html#Class_Comments)
and the [Doxygen manual](https://www.stack.nl/~dimitri/doxygen/manual/docblocks.html).