Delete Makefile & add CMakeLists.txt; add new vector struct.
Switching to cmake from simply make is neccessary for adding
a C++ Testing Framework to a project. All old include files are removed to be
replaced with a simpler aproach (less classes, more structs). Vectors are
defined as structs of x, y, and z. Functions that operate on vectors are also
declared and defined.