Add header files for atoms, molecules, bonds, observables, and others.
Atoms, molecules and bonds are interrelated structs. Observables can be of two
types: vector and scalar, with corresponding struct. There is are interface
classes for an observable container and an integrator; and partially completed
classes for a simulation and a Lennard-Jones potential. All of this will need
to be implemented.