Change definition of bond to remove atom references, do indices instead.
Since it does not even make sense for bonds to exist outside of atoms, each bond
now contains two atom indices, not atom pointers. These indices refer to the
indices of atoms in the respective atom vector containted in a given molecule
(the atom vector is never to be sorted). While this design is obviously bad,
changing the implementation --- making atoms, bonds and molecules classes, not
structs, with their own private members --- would take a very long time.