Implement state struct, improve atom / molecule / bond input and output.
A state contains time, number of molecules, and a vector of molecules. It can
be output to a string and written to a file. There is an untested implementation
of conversion of non-verbose string reperesentations of moleciules to molecule
structs; it is a good idea to test it together with the conversion of non-verbose
representation of states to the corresponding structs. This will only require
writing one more function that converts a non-verbose string representation of
state to the corresponding struct.