Finish basic tool development: simulation, force loop and Verlet intergrator.
- Give update_force functionality to ForceUpdater class. In addition,
an instance of ForceUpdater is able to update potential energy / virial
accumulators of correspodning observables, if it is given the corresponding
pointers. Inta- and inter- molecular force loops are separated, with a possibility
to provide distinct potentials to each in the future.

- Make Integrator a base class which Verlet integrator inherits. Any integrator
is required to have a ForceUpdater as a data member.

- Make Rattle integrator (unfinished) inherit from the Verlet integrator.

- Add rudimentary python function for plotting CSV files with pandas and matplotlib.

- Implement simulation class, make it a base class which specific simulations inherit.
Since difference simulations will have different parameters, states and corresponding
initializations, this should be a good move.